# Improve performance of `rem_euclid()` for signed integers