@thomcc 2022-11-04
I was not able to reproduce these benchmark numbers locally. I actually found your version consistently slightly slower. Note that `wrapping_abs` [contains a branch](https://github.com/rust-lang/rust/blob/master/library/core/src/num/int_macros.rs#L1417).

In your benchmark, the outer loop has the same value every iteration -- when checking the version that is hard on the branch predictor, you should be sure to shuffle the groups of `(a, b)`, for example:

```rs
fn samples() -> Vec<(i32, i32)> {
	let n = 8000;
    let mut all = (-n..=n)
        .flat_map(|n| (-n..=n).map(move |d| (n, d)))
        .filter(|(n, d)| n.checked_rem(*d).is_some())
        .collect::<Vec<_>>();
    let mut r = SmallRng::from_seed(SEED);
    all.shuffle(&mut r);
    all
}
```

@rustbot author