@thomcc 2022-11-04
My benchmark seems around ~5% slower on aarch64, but has smaller code size. On x86_64 it's 10% faster with much cleaner/smaller assembly and one fewer potentially-unpredictable branch. I think we should probably take it.

Please update the comment, though.

I'll trigger a perf run too, just in case this happens to be used in some weird data structure in the compiler.

@bors try @rust-timer queue